removed old react-router types. Updated redirects to happen immediately
test were failing after merging analytics updates.
