+++ package.json
@@ -74,7 +74,6 @@
     "@types/react": "^17.0.39",
     "@types/react-dom": "^17.0.9",
     "@types/react-redux": "^7.1.22",
-    "@types/react-router-dom": "^5.1.8",
     "@typescript-eslint/eslint-plugin": "^5.59.9",
     "@typescript-eslint/parser": "^5.59.9",
     "babel-jest": "^29.3.1",
